Move SourceLocation testing files to the `Fixie.Tests.Internal` namespace, to match the new location of the code under test.
