+++ src/Fixie.Tests/Internal/SourceLocationSamples.cs
@@ -1,4 +1,4 @@
-﻿namespace Fixie.Tests.TestAdapter;
+﻿namespace Fixie.Tests.Internal;
 
 // Avoid changes in this file that would move the well-known line numbers indicated in comments.
 // Otherwise, these comments and the associated assertions would need to be updated together
